+++ PV2D.h
@@ -12,23 +12,23 @@
 
 class PV2D{
     private:
-    	GLfloat x;
-    	GLfloat y;
-        GLfloat h;
+    	GLdouble x;
+    	GLdouble y;
+        GLdouble h;
     public:
         PV2D();
-        PV2D(GLfloat x, GLfloat y, GLfloat h);
-        GLfloat getX();
-        GLfloat getY();
-        GLfloat getH();
-        void setX(GLfloat x);
-        void setY(GLfloat y);
-        void setH(GLfloat h);
-        GLfloat vectorModule();
+        PV2D(GLdouble x, GLdouble y, GLdouble h);
+        GLdouble getX();
+        GLdouble getY();
+        GLdouble getH();
+        void setX(GLdouble x);
+        void setY(GLdouble y);
+        void setH(GLdouble h);
+        GLdouble vectorModule();
         PV2D normalizeVector();
-        GLfloat dot(PV2D w);
+        GLdouble dot(PV2D w);
         PV2D normalVector();
-        std::vector<GLfloat> vectorDecomposition(PV2D v, PV2D n);
+        std::vector<GLdouble> vectorDecomposition(PV2D v, PV2D n);
 };
 
 #endif
